+++ config/inflections.rb
@@ -7,17 +7,21 @@ HCloud.loader.inflector.inflect(
   "floating_ip_prices" => "FloatingIPPrices",
   "floating_ips" => "FloatingIPs",
   "hcloud" => "HCloud",
+  "health_check_http" => "HealthCheckHTTP",
   "http" => "HTTP",
   "ip_not_available" => "IPNotAvailable",
   "ipv4" => "IPv4",
   "ipv6" => "IPv6",
   "iso" => "ISO",
   "iso_type" => "ISOType",
+  "service_http" => "ServiceHTTP",
   "ssh_key" => "SSHKey",
+  "target_ip" => "TargetIP",
 )
 
 ActiveSupport::Inflector.inflections(:en) do |inflect|
   inflect.acronym "DNS"
+  inflect.acronym "HTTP"
   inflect.acronym "IP"
   inflect.acronym "ISO"
   inflect.acronym "JSON"