+++ lib/hcloud/entities/health_status.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module HCloud
+  class HealthStatus < Entity
+    attribute :listen_port, :integer
+    attribute :status
+  end
+end