+++ lib/hcloud/entities/label_selector.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module HCloud
+  class LabelSelector < Entity
+    attribute :selector
+  end
+end