+++ lib/hcloud/entities/load_balancer_private_net.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module HCloud
+  class LoadBalancerPrivateNet < Entity
+    attribute :ip
+    attribute :network
+  end
+end