+++ lib/hcloud/entities/load_balancer_public_net.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module HCloud
+  class LoadBalancerPublicNet < Entity
+    attribute :enabled, :boolean
+    attribute :ipv4, :dns_pointer
+    attribute :ipv6, :dns_pointer
+
+    alias enabled? enabled
+  end
+end