+++ lib/hcloud/entities/target_ip.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module HCloud
+  class TargetIP < Entity
+    attribute :ip
+  end
+end